+++ Source/Core/Visualization/Renderer/ScatterPlotRenderer.h
@@ -3,18 +3,8 @@
  *  @file   ScatterPlotRenderer.h
  *  @author Naohisa Sakamoto
  */
-/*----------------------------------------------------------------------------
- *
- *  Copyright (c) Visualization Laboratory, Kyoto University.
- *  All rights reserved.
- *  See http://www.viz.media.kyoto-u.ac.jp/kvs/copyright/ for details.
- *
- *  $Id: ScatterPlotRenderer.h 1797 2014-08-04 01:36:37Z naohisa.sakamoto@gmail.com $
- */
 /*****************************************************************************/
-#ifndef KVS__SCATTER_PLOT_RENDERER_H_INCLUDE
-#define KVS__SCATTER_PLOT_RENDERER_H_INCLUDE
-
+#pragma once
 #include <kvs/RendererBase>
 #include <kvs/Module>
 #include <kvs/RGBColor>
@@ -45,19 +35,18 @@ private:
     int m_bottom_margin; ///< bottom margin
     int m_left_margin; ///< left margin
     int m_right_margin; ///< right margin
-    bool m_has_point_color; ///< check flag whether the point color is specified
     kvs::RGBColor m_point_color; ///< point color
     kvs::Real32 m_point_opacity; ///< point opacity
     kvs::Real32 m_point_size; ///< point size
-    kvs::RGBColor m_edge_color;
-    kvs::Real32 m_edge_opacity;
-    kvs::Real32 m_edge_width;
-    kvs::RGBColor m_polyline_color;
-    kvs::Real32 m_polyline_opacity;
-    kvs::Real32 m_polyline_width;
+    kvs::RGBColor m_edge_color; ///< edge color
+    kvs::Real32 m_edge_opacity; ///< edge opacity
+    kvs::Real32 m_edge_width; ///< edge width
+    kvs::RGBColor m_polyline_color; ///< polyline color
+    kvs::Real32 m_polyline_opacity; ///< polyline opacity
+    kvs::Real32 m_polyline_width; ///< polyline width
     kvs::ColorMap m_color_map; ///< color map
     kvs::RGBAColor m_background_color; ///< background color
-    bool m_enable_polyline;
+    bool m_enable_polyline; ///< flag for drawing polyline
     kvs::Painter m_painter; ///< painter
 
 public:
@@ -67,7 +56,7 @@ public:
     void setBottomMargin( const int margin ) { m_bottom_margin = margin; }
     void setLeftMargin( const int margin ) { m_left_margin = margin; }
     void setRightMargin( const int margin ) { m_right_margin = margin; }
-    void setPointColor( const kvs::RGBColor color ) { m_has_point_color = true; m_point_color = color; }
+    void setPointColor( const kvs::RGBColor color ) { m_point_color = color; }
     void setPointOpacity( const kvs::Real32 opacity ) { m_point_opacity = opacity; }
     void setPointSize( const kvs::Real32 size ) { m_point_size = size; }
     void setEdgeColor( const kvs::RGBColor color ) { m_edge_color = color; }
@@ -76,7 +65,7 @@ public:
     void setPolylineColor( const kvs::RGBColor color ) { m_polyline_color = color; }
     void setPolylineOpacity( const kvs::Real32 opacity ) { m_polyline_opacity = opacity; }
     void setPolylineWidth( const kvs::Real32 width ) { m_polyline_width = width; }
-    void setColorMap( const kvs::ColorMap& color_map ) { m_has_point_color = true; m_color_map = color_map; }
+    void setColorMap( const kvs::ColorMap& color_map ) { m_color_map = color_map; }
     void setBackgroundColor( const kvs::RGBAColor color ) { m_background_color = color; }
     void setEnabledPolyline( const bool enabled ) { m_enable_polyline = enabled; }
 
@@ -93,5 +82,3 @@ public:
 };
 
 } // end of namespace kvs
-
-#endif // KVS__SCATTER_PLOT_RENDERER_H_INCLUDE