+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 ### Version 2.x.x Released (202x.xx.xx)
+**Added new classes and functions**
++ kvs::ValueAxis
++ kvs::DateAxis
+
 **Added new methods**
 + kvs::Matrix{22,33,44,nm}::rank
 + kvs::BitArray::BitArray( size, flag )
++ kvs::String::ToString( src, format )
++ kvs::String::ToString( src, precision, fixed, scientific )
 
 ### Version 2.9.0 Released (2020.5.10)
 **Added new classes and functions**